+++ flight/PiOS/STM32F4xx/pios_spi.c
@@ -214,12 +214,12 @@ int32_t PIOS_SPI_SetClockSpeed(uint32_t spi_id, uint32_t spi_speed)
 	SPIPrescalerTypeDef spi_prescaler;
 
 	//The needed prescaler for desired speed
-	uint32_t desiredPrescaler=0;
+	float desiredPrescaler=0;
 
 	if(spi_dev->cfg->regs == SPI1){
-		desiredPrescaler = 84000000/spi_speed;
+		desiredPrescaler = (float)84000000/spi_speed;
 	}else{
-		desiredPrescaler = 42000000/spi_speed;
+		desiredPrescaler = (float)42000000/spi_speed;
 	}
 
 	//Choosing the existing prescaler nearest the desiredPrescaler
